requirements.txt
streamlit
google-generativeai
chromadb
sentence-transformers
Pillow
PyMuPDF
numpy
cloudconvert
torch
requests
tensorflow
pandasmatplotlib
